+++ src/secret.h
@@ -57,10 +57,20 @@ typedef struct mbedtls_sha3_context {
 
 typedef struct secretbase_context {
   int skip;
-  mbedtls_sha3_context *ctx;
+  void *ctx;
+  void (*func)(void *, const uint8_t *, size_t);
 } secretbase_context;
+
+typedef void (*const update_func)(void *, const uint8_t *, size_t);
+typedef void (*const hash_func)(update_func, void *, SEXP);
+
+void hash_object(update_func, void *, const SEXP);
+void hash_file(update_func, void *, const SEXP);
+SEXP hash_to_sexp(unsigned char *, size_t, int);
 
 SEXP secretbase_sha3(SEXP, SEXP, SEXP);
 SEXP secretbase_sha3_file(SEXP, SEXP, SEXP);
+SEXP secretbase_xxhash(SEXP, SEXP);
+SEXP secretbase_xxhash_file(SEXP, SEXP);
 
 #endif